Confirm delete post with js

+++ admin/includes/view_all_posts.php
@@ -104,7 +104,7 @@ echo "<td> $post_id</td>";
     echo "<td> $post_date</td>";
     echo "<td> <a href='../post.php?p_id={$post_id}' target='_blank'>View Post</a></td>";
     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-    echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are You Sur You Want To Delete'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
 
     echo "</tr>";
 }
